[Front] Rollback StixCoreObjectIncidentsAreaChart.js
